+++ src/ElementWriter.js
@@ -259,9 +259,12 @@ class ElementWriter extends EventEmitter {
 		}
 	}
 
-	addVector(vector, ignoreContextX, ignoreContextY, index) {
+	addVector(vector, ignoreContextX, ignoreContextY, index, forcePage) {
 		let context = this.context();
 		let page = context.getCurrentPage();
+		if (isNumber(forcePage)) {
+			page = context.pages[forcePage];
+		}
 		let position = this.getCurrentPositionOnPage();
 
 		if (page) {
@@ -322,10 +325,21 @@ class ElementWriter extends EventEmitter {
 					var v = pack(item.item);
 
 					offsetVector(v, useBlockXOffset ? (block.xOffset || 0) : ctx.x, useBlockYOffset ? (block.yOffset || 0) : ctx.y);
-					page.items.push({
-						type: 'vector',
-						item: v
-					});
+					if (v._isFillColorFromUnbreakable) {
+						// If the item is a fillColor from an unbreakable block
+						// We have to add it at the beginning of the items body array of the page
+						delete v._isFillColorFromUnbreakable;
+						const endOfBackgroundItemsIndex = ctx.backgroundLength[ctx.page];
+						page.items.splice(endOfBackgroundItemsIndex, 0, {
+							type: 'vector',
+							item: v
+						});
+					} else {
+						page.items.push({
+							type: 'vector',
+							item: v
+						});
+					}
 					break;
 
 				case 'image':